Added base logic for square finding, and gitignore

+++ squares-api-exercise/Models/Coordinate.cs
@@ -3,7 +3,7 @@
     public class Coordinate
     {
         public Guid Id { get; set; }
-        public double X { get; set; }
-        public double Y { get; set; }
+        public int X { get; set; }
+        public int Y { get; set; }
     }
 }
